move Classes only used by ExtendedTranceiver
