+++ spinnman/extended/__init__.py
@@ -1,1 +1,22 @@
+# Copyright (c) 2023 The University of Manchester
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     https://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+from .bmp_set_led import BMPSetLed
 from .de_alloc_sdram_process import DeAllocSDRAMProcess
+from .read_adc import ReadADC
+from .set_led import SetLED
+from .write_memory_flood_process import WriteMemoryFloodProcess
+
+__all__ = ["BMPSetLed", "DeAllocSDRAMProcess", "ReadADC", "SetLED",
+           "WriteMemoryFloodProcess"]
