+++ spinnman/extended/bmp_set_led.py
@@ -17,7 +17,7 @@ from spinnman.messages.scp import SCPRequestHeader
 from spinnman.messages.scp.abstract_messages import (
     AbstractSCPRequest, BMPRequest)
 from spinnman.messages.scp.enums import SCPCommand
-from .check_ok_response import CheckOKResponse
+from spinnman.messages.scp.impl.check_ok_response import CheckOKResponse
 
 
 class BMPSetLed(BMPRequest):
