+++ spinnman/extended/extended_transceiver.py
@@ -14,63 +14,32 @@
 
 # pylint: disable=too-many-arguments
 import io
-import os
-import random
 import struct
-from threading import Condition, RLock
-from collections import defaultdict
-from contextlib import contextmanager, suppress
 import logging
-import socket
 import time
-from spinn_utilities.config_holder import get_config_bool
 from spinn_utilities.log import FormatAdapter
 from spinn_utilities.logger_utils import warn_once
 from spinn_machine import CoreSubsets
 from spinnman.transceiver import Transceiver
-from spinnman.constants import (
-    BMP_POST_POWER_ON_SLEEP_TIME, BMP_POWER_ON_TIMEOUT, BMP_TIMEOUT,
-    CPU_MAX_USER, CPU_USER_OFFSET, CPU_USER_START_ADDRESS,
-    IPTAG_TIME_OUT_WAIT_TIMES, SCP_SCAMP_PORT, SYSTEM_VARIABLE_BASE_ADDRESS,
-    UDP_BOOT_CONNECTION_DEFAULT_PORT, NO_ROUTER_DIAGNOSTIC_FILTERS,
-    ROUTER_REGISTER_BASE_ADDRESS, ROUTER_DEFAULT_FILTERS_MAX_POSITION,
-    ROUTER_FILTER_CONTROLS_OFFSET, ROUTER_DIAGNOSTIC_FILTER_SIZE, N_RETRIES,
-    BOOT_RETRIES)
+from spinnman.constants import (ROUTER_REGISTER_BASE_ADDRESS,
+    ROUTER_FILTER_CONTROLS_OFFSET, ROUTER_DIAGNOSTIC_FILTER_SIZE)
 from spinnman.data import SpiNNManDataView
-from spinnman.exceptions import (
-    SpinnmanInvalidParameterException, SpinnmanException, SpinnmanIOException,
-    SpinnmanTimeoutException, SpinnmanGenericProcessException,
-    SpinnmanUnexpectedResponseCodeException,
-    SpiNNManCoresNotInStateException)
-from spinnman.model import DiagnosticFilter, MachineDimensions
-from spinnman.model.enums import (
-    CPUState, SDP_PORTS, SDP_RUNNING_MESSAGE_CODES)
+from spinnman.exceptions import SpinnmanException
+from spinnman.extended import (
+    BMPSetLed, DeAllocSDRAMProcess, ReadADC, SetLED, WriteMemoryFloodProcess)
+from spinnman.model import DiagnosticFilter
+from spinnman.model.enums import CPUState
 from spinnman.messages.scp.enums import Signal
-from spinnman.messages.scp.impl.get_chip_info import GetChipInfo
-from spinnman.messages.sdp import SDPFlag, SDPHeader, SDPMessage
-from spinnman.messages.spinnaker_boot import (
-    SystemVariableDefinition, SpinnakerBootMessages)
-from spinnman.messages.scp.enums import PowerCommand
 from spinnman.messages.scp.abstract_messages import AbstractSCPRequest
 from spinnman.messages.scp.impl import (
-    BMPSetLed, BMPGetVersion, SetPower, ReadADC, ReadFPGARegister,
-    WriteFPGARegister, IPTagSetTTO, ReverseIPTagSet, ReadMemory,
-    CountState, WriteMemory, SetLED, ApplicationRun, SendSignal, AppStop,
-    IPTagSet, IPTagClear, RouterClear, DoSync)
+    ReadMemory, ApplicationRun)
 from spinnman.connections.udp_packet_connections import (
     BMPConnection, BootConnection, SCAMPConnection)
 from spinnman.processes import (
-    DeAllocSDRAMProcess, GetMachineProcess, GetVersionProcess,
-    MallocSDRAMProcess, WriteMemoryProcess, ReadMemoryProcess,
-    GetCPUInfoProcess, GetExcludeCPUInfoProcess, GetIncludeCPUInfoProcess,
-    ReadIOBufProcess, ApplicationRunProcess, GetHeapProcess,
-    LoadFixedRouteRoutingEntryProcess, FixedConnectionSelector,
-    ReadFixedRouteRoutingEntryProcess, WriteMemoryFloodProcess,
-    LoadMultiCastRoutesProcess, GetTagsProcess, GetMultiCastRoutesProcess,
-    SendSingleCommandProcess, ReadRouterDiagnosticsProcess,
-    MostDirectConnectionSelector, ApplicationCopyRunProcess)
+    GetHeapProcess, ReadMemoryProcess, SendSingleCommandProcess,
+    WriteMemoryProcess)
 from spinnman.utilities.utility_functions import (
-    get_vcpu_address, work_out_bmp_from_machine_details)
+    work_out_bmp_from_machine_details)
 
 logger = FormatAdapter(logging.getLogger(__name__))
 
@@ -536,6 +505,73 @@ class ExtendedTransceiver(Transceiver):
             logger.info(self.__where_is_xy(x, y))
             raise
 
+    def write_memory_flood(
+            self, base_address, data, n_bytes=None, offset=0,
+            is_filename=False):
+        """
+        Write to the SDRAM of all chips.
+
+        :param int base_address:
+            The address in SDRAM where the region of memory is to be written
+        :param data:
+            The data that is to be written.  Should be one of the following:
+
+            * An instance of RawIOBase
+            * A byte-string
+            * A single integer
+            * A file name of a file to read (in which case `is_filename`
+              should be set to True)
+        :type data:
+            ~io.RawIOBase or bytes or bytearray or int or str
+        :param int n_bytes:
+            The amount of data to be written in bytes.  If not specified:
+
+            * If `data` is an RawIOBase, an error is raised
+            * If `data` is a bytearray or bytes, the length of the bytearray
+              will be used
+            * If `data` is an int, 4 will be used
+            * If `data` is a str, the size of the file will be used
+        :param int offset:
+            The offset where the valid data starts; if `data` is
+            an int, then the offset will be ignored and 0 is used.
+        :param bool is_filename:
+            True if `data` should be interpreted as a file name
+        :raise SpinnmanIOException:
+            * If there is an error communicating with the board
+            * If there is an error reading the executable
+        :raise SpinnmanInvalidPacketException:
+            If a packet is received that is not in the valid format
+        :raise SpinnmanInvalidParameterException:
+            * If one of the specified chips is not valid
+            * If `app_id` is an invalid application ID
+            * If a packet is received that has invalid parameters
+        :raise SpinnmanUnexpectedResponseCodeException:
+            If a response indicates an error during the exchange
+        """
+        process = WriteMemoryFloodProcess(self._scamp_connection_selector)
+        # Ensure only one flood fill occurs at any one time
+        with self._flood_write_lock:
+            # Start the flood fill
+            nearest_neighbour_id = self._get_next_nearest_neighbour_id()
+            if isinstance(data, io.RawIOBase):
+                process.write_memory_from_reader(
+                    nearest_neighbour_id, base_address, data, n_bytes)
+            elif isinstance(data, str) and is_filename:
+                if n_bytes is None:
+                    n_bytes = os.stat(data).st_size
+                with open(data, "rb") as reader:
+                    process.write_memory_from_reader(
+                        nearest_neighbour_id, base_address, reader, n_bytes)
+            elif isinstance(data, int):
+                data_to_write = _ONE_WORD.pack(data)
+                process.write_memory_from_bytearray(
+                    nearest_neighbour_id, base_address, data_to_write, 0)
+            else:
+                if n_bytes is None:
+                    n_bytes = len(data)
+                process.write_memory_from_bytearray(
+                    nearest_neighbour_id, base_address, data, offset, n_bytes)
+
     def set_leds(self, x, y, cpu, led_states):
         """
         Set LED states.
@@ -686,3 +722,21 @@ class ExtendedTransceiver(Transceiver):
         warn_once(logger, "The bmp_connection property is deprecated and "
                   "likely to be removed.")
         return self._bmp_connection_selectors
+
+    def get_heap(self, x, y, heap=SystemVariableDefinition.sdram_heap_address):
+        """
+        Get the contents of the given heap on a given chip.
+
+        :param int x: The x-coordinate of the chip
+        :param int y: The y-coordinate of the chip
+        :param SystemVariableDefinition heap:
+            The SystemVariableDefinition which is the heap to read
+        :rtype: list(HeapElement)
+        """
+        try:
+            process = GetHeapProcess(self._scamp_connection_selector)
+            return process.get_heap((x, y), heap)
+        except Exception:
+            logger.info(self.__where_is_xy(x, y))
+            raise
+
