+++ spinnman/extended/read_adc.py
