+++ spinnman/extended/set_led.py
@@ -17,7 +17,7 @@ from spinnman.messages.scp import SCPRequestHeader
 from spinnman.messages.scp.abstract_messages import AbstractSCPRequest
 from spinnman.messages.scp.enums import SCPCommand
 from spinnman.messages.sdp import SDPFlag, SDPHeader
-from .check_ok_response import CheckOKResponse
+from spinnman.messages.scp.impl.check_ok_response import CheckOKResponse
 
 
 class SetLED(AbstractSCPRequest):
