+++ spinnman/extended/write_memory_flood_process.py
@@ -15,7 +15,7 @@
 import math
 from spinnman.messages.scp.impl import (
     FloodFillEnd, FloodFillStart, FloodFillData)
-from .abstract_multi_connection_process import AbstractMultiConnectionProcess
+from spinnman.processes.abstract_multi_connection_process import AbstractMultiConnectionProcess
 from spinnman.constants import UDP_MESSAGE_MAX_SIZE
