+++ spinnman/messages/scp/impl/__init__.py
@@ -16,7 +16,6 @@ from .app_stop import AppStop
 from .app_copy_run import AppCopyRun
 from .application_run import ApplicationRun
 from .bmp_get_version import BMPGetVersion
-from .bmp_set_led import BMPSetLed
 from .check_ok_response import CheckOKResponse
 from .count_state import CountState
 from .do_sync import DoSync
@@ -31,7 +30,6 @@ from .iptag_get import IPTagGet
 from .iptag_get_info import IPTagGetInfo
 from .iptag_set import IPTagSet
 from .iptag_set_tto import IPTagSetTTO
-from .read_adc import ReadADC
 from .read_fpga_register import ReadFPGARegister
 from .read_link import ReadLink
 from .read_memory import ReadMemory
@@ -42,7 +40,6 @@ from .router_init import RouterInit
 from .sdram_alloc import SDRAMAlloc
 from .sdram_de_alloc import SDRAMDeAlloc
 from .send_signal import SendSignal
-from .set_led import SetLED
 from .set_power import SetPower
 from .write_fpga_register import WriteFPGARegister
 from .write_link import WriteLink
@@ -51,15 +48,13 @@ from .fixed_route_init import FixedRouteInit
 from .fixed_route_read import FixedRouteRead
 
 __all__ = ["AppStop", "ApplicationRun", "AppCopyRun",
-           "BMPSetLed", "BMPGetVersion",
-           "CheckOKResponse", "GetChipInfo", "CountState",
+           "BMPGetVersion", "CheckOKResponse", "GetChipInfo", "CountState",
            "DoSync",
            "FloodFillData", "FillRequest",
            "FloodFillEnd", "FloodFillStart",
            "IPTagClear", "IPTagGet", "IPTagGetInfo",
-           "IPTagSet", "IPTagSetTTO", "SetLED",
-           "SetPower", "ReadADC",
-           "ReadFPGARegister", "ReadLink",
+           "IPTagSet", "IPTagSetTTO",
+           "SetPower", "ReadFPGARegister", "ReadLink",
            "ReadMemory", "ReverseIPTagSet",
            "RouterAlloc", "RouterClear",
            "RouterInit", "SDRAMAlloc",
