+++ spinnman/processes/__init__.py
@@ -39,7 +39,6 @@ from .read_memory_process import ReadMemoryProcess
 from .read_router_diagnostics_process import ReadRouterDiagnosticsProcess
 from .round_robin_connection_selector import RoundRobinConnectionSelector
 from .send_single_command_process import SendSingleCommandProcess
-from .write_memory_flood_process import WriteMemoryFloodProcess
 from .write_memory_process import WriteMemoryProcess
 
 __all__ = ["AbstractMultiConnectionProcessConnectionSelector",
@@ -47,7 +46,7 @@ __all__ = ["AbstractMultiConnectionProcessConnectionSelector",
            "RoundRobinConnectionSelector",
            "AbstractMultiConnectionProcess",
            "ApplicationRunProcess", "ApplicationCopyRunProcess",
-           "DeAllocSDRAMProcess", "GetCPUInfoProcess",
+           "GetCPUInfoProcess",
            "GetExcludeCPUInfoProcess", "GetIncludeCPUInfoProcess",
            "GetHeapProcess",
            "GetMachineProcess", "GetMultiCastRoutesProcess", "GetTagsProcess",
@@ -55,5 +54,4 @@ __all__ = ["AbstractMultiConnectionProcessConnectionSelector",
            "LoadMultiCastRoutesProcess", "MallocSDRAMProcess",
            "ReadFixedRouteRoutingEntryProcess", "ReadIOBufProcess",
            "ReadMemoryProcess", "ReadRouterDiagnosticsProcess",
-           "SendSingleCommandProcess", "WriteMemoryFloodProcess",
-           "WriteMemoryProcess"]
+           "SendSingleCommandProcess", "WriteMemoryProcess"]
